// journalist-client/src/model/bindings/Group.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GroupId } from "./GroupId";
import type { GroupMessage } from "./GroupMessage";
import type { SentinelIdentity } from "./SentinelIdentity";

/**
 * The object passed from Sentinel's backend to the frontend with group information,
 * group members, and all group messages.
 */
export type Group = {
  id: GroupId;
  display_name: string;
  description: string;
  members: Array<SentinelIdentity>;
  messages: Array<GroupMessage>;
};